:truck:  Rename Contact component to SocialNews
