+++ src/components/SocialNews/index.js
@@ -1,0 +1,1 @@
+export { default } from './SocialNews';
